third_party/blink/web_tests/platform/generic/fast/sub-pixel/float-with-right-margin-zoom-expected.txt

PASS elements[0].getBoundingClientRect().top is elements[1].getBoundingClientRect().top
PASS successfullyParsed is true

TEST COMPLETE


Both elements below should be on the same line.
